FastAPI/requirements.txt
fastapi
uvicorn
huggingface_hub
python-dotenv
Pillow
